+++ src/Shared/TagWriter/TagWriter.cs
@@ -10,6 +10,11 @@ internal abstract class TagWriter<TTagState, TArrayState>
     where TTagState : notnull
     where TArrayState : notnull
 {
+    internal const int MaxRecursionDepth = 3; // TODO https://github.com/open-telemetry/semantic-conventions/issues/3648
+
+    [ThreadStatic]
+    private static int recursionDepth;
+
     private readonly ArrayTagWriter<TArrayState> arrayWriter;
 
     protected TagWriter(
@@ -86,6 +91,40 @@ internal abstract class TagWriter<TTagState, TArrayState>
             case double d:
                 this.WriteFloatingPointTag(ref state, key, d);
                 break;
+            case IEnumerable<KeyValuePair<string, object?>> kvList:
+                try
+                {
+                    if (recursionDepth >= MaxRecursionDepth)
+                    {
+                        var stringValue = Convert.ToString(value, CultureInfo.InvariantCulture);
+                        this.WriteStringTag(
+                            ref state,
+                            key,
+                            TruncateString(stringValue.AsSpan(), tagValueMaxLength));
+
+                        break;
+                    }
+                    else
+                    {
+                        recursionDepth++;
+                    }
+
+                    this.WriteKvListTag(ref state, key, kvList, tagValueMaxLength);
+                    recursionDepth--;
+                }
+                catch (Exception ex) when (ex is IndexOutOfRangeException or ArgumentException)
+                {
+                    recursionDepth = 0;
+                    throw;
+                }
+                catch
+                {
+                    recursionDepth--;
+                    return this.LogUnsupportedTagTypeAndReturnFalse(key, value);
+                }
+
+                break;
+
             case Array array:
                 if (value.GetType() == typeof(byte[]) && this.TryWriteByteArrayTag(ref state, key, ((byte[])value).AsSpan()))
                 {
@@ -160,9 +199,16 @@ internal abstract class TagWriter<TTagState, TArrayState>
 
     protected abstract void WriteArrayTag(ref TTagState state, string key, ref TArrayState value);
 
+    protected abstract void WriteKvListTag(ref TTagState state, string key, IEnumerable<KeyValuePair<string, object?>> kvList, int? tagValueMaxLength);
+
     protected abstract void OnUnsupportedTagDropped(
         string tagKey,
         string tagValueTypeFullName);
+
+    private static ReadOnlySpan<char> TruncateString(ReadOnlySpan<char> value, int? maxLength)
+        => maxLength.HasValue && value.Length > maxLength
+           ? value.Slice(0, maxLength.Value)
+           : value;
 
     private void WriteCharTag(ref TTagState state, string key, char value)
     {