+++ src/OpenTelemetry.Exporter.Zipkin/CHANGELOG.md
@@ -6,6 +6,10 @@ Notes](../../RELEASENOTES.md).
 
 ## Unreleased
 
+* Added support for serializing attribute values that are key/value lists
+  (`IEnumerable<KeyValuePair<string, object?>>`). No change to existing behavior.
+  ([#7015](https://github.com/open-telemetry/opentelemetry-dotnet/pull/7015))
+
 ## 1.17.0
 
 Released 2026-Jul-16